+++ data-frame/columns.go
@@ -2,23 +2,22 @@ package df
 
 import (
 	"errors"
-	"fmt"
 	"reflect"
-	"strings"
 )
 
 // column represents a column inside a DataFrame
 type column struct {
-	cells    cells
+	cells    Cells
 	colType  string
 	colName  string
 	numChars int
+	empty    Cell
 }
 
 type columns []column
 
 // newCol is the constructor for a new Column with the given colName and elements
-func newCol(colName string, elements cells) (*column, error) {
+func newCol(colName string, elements Cells) (*column, error) {
 	col := column{
 		colName: colName,
 	}
@@ -28,21 +27,6 @@ func newCol(colName string, elements cells) (*column, error) {
 	}
 
 	return &col, nil
-}
-
-// Implementing the Stringer interface for Column
-func (col column) String() string {
-	strArray := []string{}
-
-	for i := 0; i < len(col.cells); i++ {
-		strArray = append(strArray, col.cells[i].String())
-	}
-
-	return fmt.Sprintln(
-		col.colName,
-		"(", col.colType, "):\n",
-		strings.Join(strArray, "\n "),
-	)
 }
 
 func parseColumn(col column, t string) (*column, error) {
@@ -80,12 +64,13 @@ func (col *column) recountNumChars() {
 }
 
 // Append will add a value or values to a column
-func (col column) append(values ...cell) (column, error) {
+func (col column) append(values ...Cell) (column, error) {
 	if len(values) == 0 {
 		col.recountNumChars()
 		return col, nil
 	}
 
+	col.empty = values[0].NA()
 	for _, v := range values {
 		t := reflect.TypeOf(v).String()
 		if col.colType == "" {
@@ -106,7 +91,7 @@ func (col column) append(values ...cell) (column, error) {
 
 func (col column) hasNa() bool {
 	for _, v := range col.cells {
-		if v.NA() {
+		if v.IsNA() {
 			return true
 		}
 	}
@@ -116,7 +101,7 @@ func (col column) hasNa() bool {
 func (col column) na() []bool {
 	naArray := make([]bool, len(col.cells))
 	for k, v := range col.cells {
-		if v.NA() {
+		if v.IsNA() {
 			naArray[k] = true
 		} else {
 			naArray[k] = false